Correciones varias v6

+++ SistemaBronx.Application/Models/ViewModels/VMPedido.cs
@@ -13,7 +13,7 @@ namespace SistemaBronx.Application.Models.ViewModels
 
         public decimal? SubTotal { get; set; }
 
-        public int? PorcDescuento { get; set; }
+        public decimal? PorcDescuento { get; set; }
 
         public decimal? ImporteAbonado { get; set; }
         public decimal? ImporteTotal { get; set; }
