+++ SistemaBronx.Models/PedidosDetalle.cs
@@ -19,9 +19,9 @@ public partial class PedidosDetalle
 
     public decimal? CostoUnitario { get; set; }
 
-    public int? PorcGanancia { get; set; }
+    public decimal? PorcGanancia { get; set; }
 
-    public int? PorcIva { get; set; }
+    public decimal? PorcIva { get; set; }
     public string Producto { get; set; }
 
     public decimal? PrecioVenta { get; set; }
